Unified tree tests
