+++ modules/doobie/src/test/scala/tree/TreeMapping.scala
@@ -1,0 +1,22 @@
+// Copyright (c) 2016-2020 Association of Universities for Research in Astronomy, Inc. (AURA)
+// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
+package tree
+
+import cats.effect.Sync
+import doobie.util.transactor.Transactor
+
+import edu.gemini.grackle.Mapping
+import edu.gemini.grackle.doobie.{DoobieMappingCompanion, DoobieMonitor}
+
+import utils.{DatabaseSuite, DoobieTestMapping}
+import grackle.test.SqlTreeSpec
+
+object TreeMapping extends DoobieMappingCompanion {
+  def mkMapping[F[_]: Sync](transactor: Transactor[F], monitor: DoobieMonitor[F]): Mapping[F] =
+    new DoobieTestMapping[F](transactor, monitor) with SqlTreeMapping[F]
+}
+
+final class TreeSpec extends DatabaseSuite with SqlTreeSpec {
+  lazy val mapping = TreeMapping.mkMapping(xa)
+}
