+++ modules/skunk/src/test/scala/tree/TreeMapping.scala
@@ -1,0 +1,22 @@
+// Copyright (c) 2016-2020 Association of Universities for Research in Astronomy, Inc. (AURA)
+// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
+package tree
+
+import cats.effect.{Resource, Sync}
+import skunk.Session
+
+import edu.gemini.grackle.Mapping
+import edu.gemini.grackle.skunk.{SkunkMappingCompanion, SkunkMonitor}
+
+import utils.{DatabaseSuite, SkunkTestMapping}
+import grackle.test.SqlTreeSpec
+
+object TreeMapping extends SkunkMappingCompanion {
+  def mkMapping[F[_]: Sync](pool: Resource[F, Session[F]], monitor: SkunkMonitor[F]): Mapping[F] =
+    new SkunkTestMapping[F](pool, monitor) with SqlTreeMapping[F]
+}
+
+final class TreeSpec extends DatabaseSuite with SqlTreeSpec {
+  lazy val mapping = TreeMapping.mkMapping(pool)
+}
